+++ resources/views/user/edit.blade.php
@@ -14,6 +14,7 @@
     <h2 class="mt-5 mb-3">ユーザ情報を編集する</h2>
     <form method="POST" action="{{ route('user.update',$user->id) }}">
         @csrf
+        @method('PUT')
         <input type="hidden" name="id" value="{{ $user->id }}" />
         <div class="form-group">
             <label for="name">ユーザ名</label>
