+++ resources/views/users/show.blade.php
@@ -10,7 +10,7 @@
                     <img class="rounded-circle img-fluid" src="{{ Gravatar::src($user->email, 330) }}" alt="ユーザのアバター画像">
                         @if (Auth::check() && Auth::id() === $user->id)
                             <div class="mt-3">
-                                <a href="" class="btn btn-primary btn-block">ユーザ情報の編集</a>
+                                <a href="{{ route('user.edit', $user->id) }}" class="btn btn-primary btn-block">ユーザ情報の編集</a>
                             </div>
                         @endif
                 </div>
